fn_80127F24: s8 loop counter restores test-first loops + extsb (88.9->93.0%)
Both i-loops read the counter as (s8) (extsb before subfic/cmpwi) and use
test-first form (initial b to loop test) in target. Declaring i as s8 instead
of s32 reproduces the extsb extension and the test-first branch structure.

(cherry picked from commit 2050bc31d8d66c7691a18d6bd93719aeb68858a7)

+++ src/main/dll/pausemenu.c
@@ -721,7 +721,7 @@ void fn_80127F24(s32 alpha)
     f32 denom;
     f32 phase;
     f32 brightness;
-    s32 i;
+    s8 i;
 
     phase = lbl_803E1F18 *
         mathSinf(lbl_803E1EC8 * (lbl_803DD748 * lbl_803E201C) /
